Call the link outcome a result rather than a preparation
The enum says what the caller gets back, not the work that produced it.
